// extensions/nodejs-engine/src/model/base-schema.ts
export interface components {
    schemas: {
// ...
        EventSelector: {
            actions?: ("CREATE" | "UPDATE" | "DELETE" | "GET" | "LIST" | "OPERATE")[];
            recordSelector?: components["schemas"]["BooleanExpression"];
            namespaces?: (string)[];
            resources?: (string)[];
            /** @description star means all */
            ids?: (string)[];
            /** @description star means all, empty means proceed */
            annotations?: {
                [key: string]: string | undefined;
            };
        };
// ...
        ExternalCall: {
            functionCall?: components["schemas"]["FunctionCall"];
            httpCall?: components["schemas"]["HttpCall"];
        };
// ...
        HttpCall: {
            uri?: string;
            method?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}